refactor: change render logic, change directory
details below:
- render winner after render racing arrow ends
- change method names from derender to erase
- simplify makeArrToString function
- move utitls.js and validation.js to utils directory
