+++ src/controllers/userController.js
@@ -1,5 +1,5 @@
-import { isNameValid, isRacingNumberValid, isUserInputExist } from './validation.js';
-import { doTrim } from './utils.js';
+import { isNameValid, isRacingNumberValid, isUserInputExist } from '../utils/validation.js';
+import { doTrim } from '../utils/utils.js';
 import { race, clearState } from './raceController.js';
 import {
   renderRacingInputForm,
